azure_ai_requirements.txt
langchain-azure-ai>=0.1.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
requests>=2.25.0 